Lot's of changes!
Restructured code and added packages.
Added HintProvider that can provide hints for a given secret word and guessed word.
Bug fix: Found a bug that was leading the rejection of the secret word out of solution set because of the incorrect count of character in mandatoyInclusions list. Have fixed it and realized that mandatoryInclusions DS can be improvised.